requirements.txt
beautifulsoup4
openai
python-dotenv
selenium
anthropic
